+++ MHLibDemos/test/tests.jl
@@ -50,7 +50,7 @@ end
     println(sol)
     num_de = 5
     method_selector = WeightedRandomMethodSelector(num_de:-1:1, 1:1)
-    alg = LNS(sol, [MHMethod("construct", construct!)],
+    alg = LNS(sol, [MHMethod("const", construct!)],
         [MHMethod("de$i", destroy!, i) for i in 1:num_de],
         [MHMethod("re", repair!)]; method_selector, titer=120)
     run!(alg)
@@ -65,7 +65,7 @@ end
     println(sol)
     num_de = 3
     method_selector = WeightedRandomMethodSelector(num_de:-1:1, 1:1)
-    alg = LNS(sol, [MHMethod("construct", construct!)],
+    alg = LNS(sol, [MHMethod("const", construct!)],
         [MHMethod("de$i", destroy!, i) for i in 1:num_de],
         [MHMethod("re", repair!)]; method_selector, titer=120)
     run!(alg)
@@ -79,7 +79,7 @@ end
     sol = MAXSATSolution(inst)
     println(sol)
     num_de = 5
-    alg = ALNS(sol, [MHMethod("construct", construct!)],
+    alg = ALNS(sol, [MHMethod("const", construct!)],
         [MHMethod("de$i", destroy!, i) for i in 1:num_de],
         [MHMethod("re", repair!)], titer=120)
     run!(alg)